requirements.txt
Flask>=2.2.0
numpy>=1.23.0
Pillow>=9.0.0
tensorflow-cpu>=2.10.0
gunicorn>=20.1.0
streamlit>=1.20.0